Work on Stacks and Queues
This folder contains 2 different implementations for both Stacks and Queues. One with Static Arrays and the other with Dynamic Arrays.

Also contains a simple Makefile to ease testing and using the code.

There won't be any code comments as of right now since this code only implements basic queries that the buffer requires to be used and the solution is quite simple.